test/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
ultralytics>=8.3,<9
pillow>=10,<12
python-multipart>=0.0.9,<1
